implement update and delete transaction functionalities in the transaction API and UI components

+++ server/controllers/transaction.controller.js
@@ -50,9 +50,9 @@ export const updateTransactions = async (req, res) => {
     
     const updateTrans = await Transaction.findByIdAndUpdate(id, updateData, {new: true})
 
-    console.log(updateTrans);
+    // console.log(updateTrans);
     
-    return res.status(200).json({success: true, updateTrans})
+    return res.json(updateTrans)
 
   } catch (error) {
     console.log(error);
@@ -64,7 +64,11 @@ export const deleteTransaction = async (req, res) => {
   try {
     const {id} = req.params
 
-    // console.log(req.params);
+    console.log(req.params);
+
+    if (!id) {
+      return res.status(400).json({ message: 'Transaction ID is required' });
+    }
     
     const trashed = await Transaction.findByIdAndDelete(id);
     console.log(trashed);
